// src/views/Voting/types.ts
export enum ProposalType {
  ALL = 'all',
  CORE = 'core',
}

export enum FilterId {
  NOW = 'now',
  SOON = 'soon',
  CLOSED = 'closed',
}

export enum BadgeType {
  CORE = 'core',
  PARTICIPATION = 'Participation',
}

export enum TransactionState {
  NONE = 'none',
  START = 'start',
  SUCCESS = 'success',
  ERROR = 'error',
}

export enum ModalState {
  VOTING = 'voting',
  CONFIRM = 'confirm',
}

export default {}